Prototyping analysis thread

+++ MeteorCaptureQt/infra/image.h
@@ -24,7 +24,7 @@ public:
      * Records the epoch time in microseconds, i.e. the time elapsed since 1970-01-01T00:00:00Z,
      * of the time the first byte of image data was captured.
      */
-    long epochTimeUs;
+    long long epochTimeUs;
